medium photos AND content files draggable, sending correctly to frontend. all go
