+++ config/routes.rb
@@ -14,13 +14,15 @@ Rails.application.routes.draw do
     end
   end
   resources :people
-  resources :archive_items, only: [:index, :create, :show, :new, :edit, :destroy, :update, :delete] do
+  resources :archive_items, only: [:index, :create, :show, :new, :edit, :destroy, :update, ] do
     collection do
       post :export_to_csv
     end
     member do
       get 'copy'
-      delete 'delete_medium_photo', 'delete_poster_image', 'delete_content_file'
+      delete 'delete_medium_photo'
+      delete 'delete_poster_image'
+      delete 'delete_content_file'
       get :create_uid_pdf
       patch :update_content_files_order
       patch :update_medium_photos_order
